slack/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
multidict==6.0.2
multitasking==0.0.10
numpy==1.22.2
pandas==1.4.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
slack-sdk==3.15.1
slackclient==1.3.2
slacker==0.14.0
urllib3==1.26.8
websocket-client==0.54.0
Werkzeug==2.0.3
wincertstore==0.2
yarl==1.7.2
yfinance==0.1.70
